+++ js/loading.js
@@ -1,0 +1,48 @@
+export const TITLES = [
+  'Домик для букашек',
+  'Три стопочки и ты дома',
+  'Девичье гнездышко',
+  'Сладкие подушечки',
+  'Домик на дереве',
+  'Сон под открытым небом',
+  'Запахи охоты и рыбалки',
+  'Особенности охоты и рыбалки',
+  'Хижина Васька',
+];
+
+export const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+export const TIMES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+export const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+export const DESCRIPTIONS = [
+  'Насладитесь умиротворением и безмятежностью в удивительном мире нового отеля с продуманной до мелочей концепцией идеального отдыха по мировым стандартам',
+  'Отель расположен на берегу Средиземного моря в окружении хвойных лесов и апельсиновых садов, обладает шикарной, просторной территорией 98 500 м2;, имеет собственный пляж протяженностью в пол километра. - Широкий выбор услуг и развлечений на открытом воздухе, позволят вам насладится отдыхом, не заходя в помещение.',
+  'Месторасположение: в поселке Гёйнюк, в 9 км от Кемера, в 49 км от аэропорта. На территории отеля зелёный сад. Транспорт - такси, автобус.   Описание номеров: 47 стандартные комнаты с видом на море, 326 стандартных комнат с видом на сад, ванна, телефон, фен, мини-бар безалкогольные напитки, ТВ с российскими каналами, балкон, ковровое покрытие, центральное кондиционирование, детская кровать. Смена белья 3 раза в неделю, полотенец через день. 12 семейных сюит комнат (48 кв.м, макс.4 чел., мин.',
+];
+
+export const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export default {TITLES, TYPES, TIMES, FEATURES, DESCRIPTIONS, PHOTOS};
